separate state_keys and state utils
