+++ alphastats/gui/pages/02_Import Data.py
@@ -12,6 +12,8 @@ from alphastats.gui.utils.import_helper import (
 from alphastats.gui.utils.options import SOFTWARE_OPTIONS
 from alphastats.gui.utils.state_keys import (
     StateKeys,
+)
+from alphastats.gui.utils.state_utils import (
     empty_session_state,
     init_session_state,
 )
