+++ alphastats/gui/pages/03_Data Overview.py
@@ -7,7 +7,8 @@ from alphastats.gui.utils.overview_helper import (
     get_intensity_distribution_unprocessed,
     get_sample_histogram_matrix,
 )
-from alphastats.gui.utils.state_keys import StateKeys, init_session_state
+from alphastats.gui.utils.state_keys import StateKeys
+from alphastats.gui.utils.state_utils import init_session_state
 from alphastats.gui.utils.ui_helper import sidebar_info
 
 st.set_page_config(layout="wide")
