+++ alphastats/gui/pages/06_LLM.py
@@ -22,6 +22,8 @@ from alphastats.gui.utils.llm_helper import (
 )
 from alphastats.gui.utils.state_keys import (
     StateKeys,
+)
+from alphastats.gui.utils.state_utils import (
     init_session_state,
 )
 from alphastats.gui.utils.ui_helper import (
