+++ alphastats/gui/pages/07_Results.py
@@ -5,6 +5,8 @@ from alphastats.gui.utils.analysis_helper import (
 )
 from alphastats.gui.utils.state_keys import (
     StateKeys,
+)
+from alphastats.gui.utils.state_utils import (
     init_session_state,
 )
 from alphastats.gui.utils.ui_helper import (
